requirements.txt
langchain
langchain_openai
mcp
python-dotenv
requests
psycopg2